requirements-azure.txt
certifi==2025.8.3
pyyaml==6.0.2
httpx[http2]==0.28.1
httpx-socks[asyncio]==0.10.0
Brotli==1.1.0
python-dateutil==2.9.0.post0
lxml==6.0.0
babel==2.17.0
flask==3.1.1
flask-babel==4.0.0
markdown-it-py==3.0.0
fasttext-predict==0.9.2.4
tomli==2.2.1; python_version < '3.11'
msgspec==0.19.0
isodate==0.7.2
azure-functions==1.18.0
packaging==24.1
requests==2.32.4
